Split perf tests
