+++ v2/tests/logger_perf/logger_perf_enums.c
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+#ifndef LOGGER_PERF_V1_H
+#define LOGGER_PERF_V1_H
+
+void logger_perf_v1_run_tests(void);
+
+#endif // LOGGER_PERF_V1_H
